Revert "Añadido función de editar empleado"
