+++ EmployEase/nbproject/private/private.properties
@@ -5,4 +5,4 @@ do.jlink=false
 javac.debug=true
 javadoc.preview=true
 jlink.strip=false
-user.properties.file=C:\\Users\\Usuario\\AppData\\Roaming\\NetBeans\\21\\build.properties
+user.properties.file=C:\\Users\\Mario\\AppData\\Roaming\\NetBeans\\21\\build.properties
