+++ EmployEase/nbproject/private/private.xml
@@ -3,7 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/D:/PracticasMicros/EmployEase/EmployEase/src/employeasepkg/ClaseEmpleado.java</file>
+            <file>file:/C:/practicas/EmployEase/EmployEase/src/FrameAñadirEmpleado.java</file>
         </group>
     </open-files>
 </project-private>
