+++ EmployEase/nbproject/project.properties
@@ -78,7 +78,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=EmployEase
-main.class=employeasepkg.FrameInicioSesion
+main.class=FrameInicioSesion
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
