--- EmployEase/src/employeasepkg/ClaseEmpleado.java
@@ -1,279 +1,0 @@
-package employeasepkg;
-
-import java.util.Date;
-
-public class ClaseEmpleado {
-
-    private Date fecha_r;
-    private int id;
-    private String nombre;
-    private String apellido;
-    private String unidad;
-    private String telefono;
-    private String correo;
-    private String nss;
-    private String curp;
-    private String rfc;
-    private String broxel;
-    private String banco;
-    private int clave_i;
-    private String direccion;
-    private Date fecha_n;
-    private String lugar_n;
-    private String sexo;
-    private String edo_civil;
-    private String no_cedula;
-    private String vacuna_covid;
-    private String puesto;
-    private String departamento;
-    private String credito_infonavit;
-    private String nombre_beneficiario;
-    private String telefono_beneficiario;
-    private String parentesco_beneficiario;
-    private String estatus;
-    private String motivo_baja;
-    private String recomendable;
-    private String motivo;
-    private String tipo;
-    private Date fecha_a;
-    private Date fecha_i;
-    private Date fecha_b;
-
-    /**
-     * @return the fecha_r
-     */
-    public Date getFecha_r() {
-        return fecha_r;
-    }
-
-    /**
-     * @return the id
-     */
-    public int getId() {
-        return id;
-    }
-
-    /**
-     * @return the nombre
-     */
-    public String getNombre() {
-        return nombre;
-    }
-
-    /**
-     * @return the apellido
-     */
-    public String getApellido() {
-        return apellido;
-    }
-
-    /**
-     * @return the unidad
-     */
-    public String getUnidad() {
-        return unidad;
-    }
-
-    /**
-     * @return the telefono
-     */
-    public String getTelefono() {
-        return telefono;
-    }
-
-    /**
-     * @return the correo
-     */
-    public String getCorreo() {
-        return correo;
-    }
-
-    /**
-     * @return the nss
-     */
-    public String getNss() {
-        return nss;
-    }
-
-    /**
-     * @return the curp
-     */
-    public String getCurp() {
-        return curp;
-    }
-
-    /**
-     * @return the rfc
-     */
-    public String getRfc() {
-        return rfc;
-    }
-
-    /**
-     * @return the broxel
-     */
-    public String getBroxel() {
-        return broxel;
-    }
-
-    /**
-     * @return the banco
-     */
-    public String getBanco() {
-        return banco;
-    }
-
-    /**
-     * @return the clave_i
-     */
-    public int getClave_i() {
-        return clave_i;
-    }
-
-    /**
-     * @return the direccion
-     */
-    public String getDireccion() {
-        return direccion;
-    }
-
-    /**
-     * @return the fecha_n
-     */
-    public Date getFecha_n() {
-        return fecha_n;
-    }
-
-    /**
-     * @return the lugar_n
-     */
-    public String getLugar_n() {
-        return lugar_n;
-    }
-
-    /**
-     * @return the sexo
-     */
-    public String getSexo() {
-        return sexo;
-    }
-
-    /**
-     * @return the edo_civil
-     */
-    public String getEdo_civil() {
-        return edo_civil;
-    }
-
-    /**
-     * @return the no_cedula
-     */
-    public String getNo_cedula() {
-        return no_cedula;
-    }
-
-    /**
-     * @return the vacuna_covid
-     */
-    public String getVacuna_covid() {
-        return vacuna_covid;
-    }
-
-    /**
-     * @return the puesto
-     */
-    public String getPuesto() {
-        return puesto;
-    }
-
-    /**
-     * @return the departamento
-     */
-    public String getDepartamento() {
-        return departamento;
-    }
-
-    /**
-     * @return the credito_infonavit
-     */
-    public String getCredito_infonavit() {
-        return credito_infonavit;
-    }
-
-    /**
-     * @return the nombre_beneficiario
-     */
-    public String getNombre_beneficiario() {
-        return nombre_beneficiario;
-    }
-
-    /**
-     * @return the telefono_beneficiario
-     */
-    public String getTelefono_beneficiario() {
-        return telefono_beneficiario;
-    }
-
-    /**
-     * @return the parentesco_beneficiario
-     */
-    public String getParentesco_beneficiario() {
-        return parentesco_beneficiario;
-    }
-
-    /**
-     * @return the estatus
-     */
-    public String getEstatus() {
-        return estatus;
-    }
-
-    /**
-     * @return the motivo_baja
-     */
-    public String getMotivo_baja() {
-        return motivo_baja;
-    }
-
-    /**
-     * @return the recomendable
-     */
-    public String getRecomendable() {
-        return recomendable;
-    }
-
-    /**
-     * @return the motivo
-     */
-    public String getMotivo() {
-        return motivo;
-    }
-
-    /**
-     * @return the tipo
-     */
-    public String getTipo() {
-        return tipo;
-    }
-
-    /**
-     * @return the fecha_a
-     */
-    public Date getFecha_a() {
-        return fecha_a;
-    }
-
-    /**
-     * @return the fecha_i
-     */
-    public Date getFecha_i() {
-        return fecha_i;
-    }
-
-    /**
-     * @return the fecha_b
-     */
-    public Date getFecha_b() {
-        return fecha_b;
-    }
-}
