--- EmployEase/src/employeasepkg/FrameAgregarEmpleado.java
@@ -1,310 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package employeasepkg;
-
-import java.awt.HeadlessException;
-import java.sql.Date;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import javax.swing.JOptionPane;
-
-/**
- *
- * @author Usuario
- */
-public class FrameAgregarEmpleado extends javax.swing.JFrame {
-
-    ConexionSQL connection = new ConexionSQL();
-
-    /**
-     * Creates new form FrameAñadirEmpleado
-     */
-    public FrameAgregarEmpleado() {
-        initComponents();
-        loadData();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        lblID = new javax.swing.JLabel();
-        btnAdd = new javax.swing.JButton();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel3 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-        lblCurrentDate = new javax.swing.JLabel();
-        jLabel4 = new javax.swing.JLabel();
-        jLabel30 = new javax.swing.JLabel();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        panelEmpleado = new employeasepkg.PanelEditarEmpleado();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        lblID.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
-        lblID.setText("0");
-
-        btnAdd.setText("Agregar registro");
-        btnAdd.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnAddActionPerformed(evt);
-            }
-        });
-
-        jLabel1.setFont(new java.awt.Font("Calibri", 1, 26)); // NOI18N
-        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel1.setText("Ingreso de Nuevo Empleado");
-
-        jLabel3.setText("Campos con * son obligatorios");
-
-        jLabel2.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
-        jLabel2.setText("N° de empleado:");
-
-        lblCurrentDate.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
-        lblCurrentDate.setText("0");
-
-        jLabel4.setFont(new java.awt.Font("Calibri", 1, 18)); // NOI18N
-        jLabel4.setText("Información personal");
-
-        jLabel30.setFont(new java.awt.Font("Calibri", 0, 18)); // NOI18N
-        jLabel30.setText("Fecha de registro:");
-
-        jScrollPane1.setViewportView(panelEmpleado);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(35, 35, 35)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jLabel4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 833, Short.MAX_VALUE)
-                            .addGroup(layout.createSequentialGroup()
-                                .addComponent(jLabel2)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(lblID, javax.swing.GroupLayout.PREFERRED_SIZE, 41, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(28, 28, 28)
-                                .addComponent(jLabel30)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(lblCurrentDate, javax.swing.GroupLayout.PREFERRED_SIZE, 177, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(btnAdd)))
-                        .addGap(31, 31, 31))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel3)
-                            .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 820, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(0, 0, Short.MAX_VALUE)))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(40, 40, 40)
-                .addComponent(jLabel1)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jLabel4)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 16, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(24, 24, 24)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel2)
-                    .addComponent(lblCurrentDate)
-                    .addComponent(jLabel30)
-                    .addComponent(lblID)
-                    .addComponent(btnAdd))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 517, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(61, Short.MAX_VALUE))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void btnAddActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAddActionPerformed
-        addRegister();
-    }//GEN-LAST:event_btnAddActionPerformed
-
-    public boolean verifyFields() {
-        String nombre = panelEmpleado.getTxtName();
-        String apellido = panelEmpleado.getTxtLastName();
-        String unidad = panelEmpleado.getTxtUnit();
-        String puesto = panelEmpleado.getTxtPosition();
-        String departamento = panelEmpleado.getTxtDepartment();
-
-        if (nombre.trim().equals("") || apellido.trim().equals("") || unidad.trim().equals("") || puesto.trim().equals("")
-                || departamento.trim().equals("")) {
-            JOptionPane.showMessageDialog(rootPane, "Faltan datos por llenar", "Error", JOptionPane.ERROR_MESSAGE);
-            return false;
-        } else {
-            return true;
-        }
-    }
-
-    private void addRegister() {
-        PreparedStatement st;
-
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        String fNacimiento = sdf.format(panelEmpleado.getCalBirthdayDate());
-        String fInicio = sdf.format(panelEmpleado.getCalBirthdayDate());
-        String fFin = sdf.format(panelEmpleado.getCalBirthdayDate());
-        String fAltaIMSS = sdf.format(panelEmpleado.getCalRegisterIMSSDate());
-
-        String fActual = lblCurrentDate.getText();
-        String nombre = panelEmpleado.getTxtName();
-        String apellido = panelEmpleado.getTxtLastName();
-        String unidad = panelEmpleado.getTxtUnit();
-        String telefono = panelEmpleado.getTxtPhone();
-        String email = panelEmpleado.getTxtEmail();
-        String nss = panelEmpleado.getTxtNSS();
-        String curp = panelEmpleado.getTxtCurp();
-        String rfc = panelEmpleado.getTxtRFC();
-        String broxel = panelEmpleado.getTxtBroxel();
-        String clave = panelEmpleado.getTxtClaveInter();
-        String direccion = panelEmpleado.getTxtDirection();
-        String edoCivil = panelEmpleado.getCboxCivilStateSelectedItem();
-        String cedula = panelEmpleado.getTxtCedula();
-        String puesto = panelEmpleado.getTxtPosition();
-        String departamento = panelEmpleado.getTxtDepartment();
-        String credito = panelEmpleado.getTxtCreditI();
-        String banco = panelEmpleado.getTxtBank();
-        String bNombre = panelEmpleado.getTxtBeneficiaryName();
-        String bTelefono = panelEmpleado.getTxtBeneficiaryPhone();
-        String bParentesco = panelEmpleado.getTxtBeneficiaryRelation();
-        String lNacimiento = panelEmpleado.getTxtBornPlace();
-        String razones = panelEmpleado.getTxtReasons();
-        String comentarios = panelEmpleado.getTxtComentaries();
-        String sexo = (panelEmpleado.isRbtnWomanSelected()) ? "Mujer" : "Hombre";
-        String vacunado = (panelEmpleado.isChkVaccinatedChecked()) ? "VACUNADO" : "REGISTRADO";
-        String estatus = (panelEmpleado.isRbtnActiveSelected()) ? "ACTIVO" : "INACTIVO";
-        String recomendable = String.valueOf(panelEmpleado.getSliderRecommended());
-        String tipo = (panelEmpleado.isRbtnEntrySelected()) ? "ALTA" : "REINGRESO";
-
-        if (verifyFields()) {
-            String query = "INSERT INTO `employease`(`fecha_ingreso`, `nombre`, `apellido`, `unidad`, "
-                    + "`telefono`, `correo`, `nss`, `curp`, `rfc`, `broxel`, `banco`, `clave_interbancaria`, `direccion`, `fecha_nacimiento`, "
-                    + "`lugar_nacimiento`, `sexo`, `estado_civil`, `no_cedula`, `vacuna_covid`, `puesto`, `departamento`, `credito_infonavit`, "
-                    + "`nombre_beneficiario`, `telefono_beneficiario`, `parentesco_beneficiario`, `estatus`, `motivo_baja`, `recomendable`, "
-                    + "`motivo`, `tipo`, `fecha_alta_imss`, `fecha_inicio`, `fecha_baja`) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-
-            try {
-                st = connection.getConectarDB().prepareStatement(query);
-                st.setDate(1, Date.valueOf(fActual));
-                st.setString(2, nombre);
-                st.setString(3, apellido);
-                st.setString(4, unidad);
-                st.setString(5, telefono);
-                st.setString(6, email);
-                st.setString(7, nss);
-                st.setString(8, curp);
-                st.setString(9, rfc);
-                st.setString(10, broxel);
-                st.setString(11, banco);
-                if (clave.trim().equals("")) {
-                    st.setInt(12, 0);
-                } else {
-                    st.setInt(12, Integer.parseInt(clave));
-                }
-                st.setString(13, direccion);
-                st.setDate(14, Date.valueOf(fNacimiento));
-                st.setString(15, lNacimiento);
-                st.setString(16, sexo);
-                st.setString(17, edoCivil);
-                st.setString(18, cedula);
-                st.setString(19, vacunado);
-                st.setString(20, puesto);
-                st.setString(21, departamento);
-                st.setString(22, credito);
-                st.setString(23, bNombre);
-                st.setString(24, bTelefono);
-                st.setString(25, bParentesco);
-                st.setString(26, estatus);
-                st.setString(27, razones);
-                st.setString(28, recomendable);
-                st.setString(29, comentarios);
-                st.setString(30, tipo);
-                st.setDate(31, Date.valueOf(fAltaIMSS));
-                st.setDate(32, Date.valueOf(fInicio));
-                st.setDate(33, Date.valueOf(fFin));
-
-                if (st.executeUpdate() != 0) {
-                    JOptionPane.showMessageDialog(rootPane, "Se añadió el empleado con éxito");
-                    //Cerrar pestaña
-                    FrameTablaEmpleados info = new FrameTablaEmpleados();
-                    info.show();
-                    this.dispose();
-                } else {
-                    JOptionPane.showMessageDialog(rootPane, "Error al añadir el empleado");
-                }
-            } catch (HeadlessException | NumberFormatException | SQLException e) {
-                System.out.println(e.toString());
-            }
-        }
-    }
-
-    private void loadData() {
-        lblCurrentDate.setText(LocalDate.now().toString()); //Fecha actual
-        //NO MOVER
-        //not null dates
-        panelEmpleado.setCaldStartDate(Date.valueOf(LocalDate.now()));
-        panelEmpleado.setCalBirthdayDate(Date.valueOf(LocalDate.now().minusYears(18)));
-    }
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException | javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(FrameAgregarEmpleado.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(() -> {
-            new FrameAgregarEmpleado().setVisible(true);
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnAdd;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel30;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JLabel lblCurrentDate;
-    private javax.swing.JLabel lblID;
-    private employeasepkg.PanelEditarEmpleado panelEmpleado;
-    // End of variables declaration//GEN-END:variables
-}
